edit sum

+++ src/com/codegym/Main.java
@@ -15,8 +15,10 @@ public class Main {
             }
         }
         System.out.println("Tính tổng cột mấy ?");
+
         int choice = sc.nextInt();
-        int sum = 0;
+
+        double sum = 0.0;
         for (int i = 0; i < 3; i++) {
             for (int j = 0; j < 3; j++) {
                 if (choice == 1) {
